# optimize DateTimeOffset.UtcNow by removing redundant verification