@adamsitnik 2020-11-30
> Need to change expected exception type in test

thanks! I could not find the failure logs due to #45357 I was not sure about changing the thrown exception (it would be a breaking change?) so I've changed the order of passed arguments.