@benaadams 2020-11-30
> it would be a breaking change?

It's a more derived exception so would be ok; though the test was doing an exact match which is why it broke